Resolve some warnings in Spring extension
