+++ extension-spring/src/test/java/test/io/smallrye/openapi/runtime/scanner/resources/GreetingPostController.java
@@ -32,7 +32,7 @@ public class GreetingPostController {
     // 2) ResponseEntity without a type specified
     @PostMapping("/greetWithResponse")
     @APIResponse(responseCode = "200", description = "OK", content = @Content(schema = @Schema(ref = "#/components/schemas/Greeting")))
-    public ResponseEntity greetWithResponse(@RequestBody Greeting greeting) {
+    public ResponseEntity<Greeting> greetWithResponse(@RequestBody Greeting greeting) {
         return ResponseEntity.ok(greeting);
     }
